refactor(react): group storybook-only helpers under stories/support
The stories/ folder mixed *.stories.tsx showcases with plain *.ts helper
files. Those helpers are storybook-only scaffolding — not library code —
so move them into a stories/support/ subfolder instead of lib/ (which is
the shipped runtime surface: cn, chart, icons, responsive-visibility).
Keeps story-height-test-utils' storybook/test dependency out of lib/ and
leaves the stories/ top level reading as pure showcases.

- Move story-height-test-utils.ts, runtime-token-values.ts, and
  typography-utilities.ts into stories/support/.
- Repoint the 12 import sites (7 component stories + 4 token-showcase
  stories) and runtime-token-values' own provider import.
