+++ packages/react/src/components/button-group/ButtonGroup.stories.tsx
@@ -8,7 +8,7 @@ import {
 } from '@tabler/icons-react';
 import { expect, within } from 'storybook/test';
 
-import { expectHeightPinned } from '../../stories/story-height-test-utils';
+import { expectHeightPinned } from '../../stories/support/story-height-test-utils';
 import { Button } from '../button';
 import {
   DropdownMenu,
